requirements.txt
pypif
peakutils
pillow
pyxrd